+++ http4k-cloudnative/src/main/kotlin/org/http4k/cloudnative/deprecatedCloudNative.kt
@@ -1,0 +1,4 @@
+package org.http4k.cloudnative
+
+@Deprecated("Renamed", ReplaceWith("RemoteRequestFailed"))
+typealias UpstreamRequestFailed = RemoteRequestFailed